Removed redundant part route url (#20)
* Got basic route to interact with Meta Live Video API

* Changed the route so it's not redundant

* Added functionality for chat to appear on the scoreboard. Need to make it so that each sport has their individual feed that they can pull from.

* Chat feature is implemented in all of the sports.

Just need to figure out a way to not have to go into the code to enter the video ID.

* Updated so that the live chat is displayed on its own page.

It also pulls the id directly from the page, so entering it is unnecessary.

Made live chat comment section scrollable, but everything else is stationary.

* Did some formatting changes

Removed a call to fetching the live comments with a parameter. Removed parameter

* just some more formatting lets to use consts